Add new modules and refactor code
Added new modules for dice rolling and block selection, and refactored the code to use ES6 modules. Also moved the logRollResult function to a separate file for better organization.
