+++ package.json
@@ -2,6 +2,7 @@
   "name": "mario-kart-node",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node --watch src/index.js"
