+++ src/game/dice.js
@@ -1,0 +1,5 @@
+async function rollDice() {
+    return Math.floor(Math.random() * 6) + 1;
+}
+
+export { rollDice };
