+++ src/utils/logger.js
@@ -1,0 +1,5 @@
+async function logRollResult(characterName, block, diceResult, attribute) {
+    console.log(`${characterName} 🎲 rolou um dado de ${block} ${diceResult} + ${attribute} = ${diceResult + attribute}`);
+}
+
+export { logRollResult };
